Make sure Optionals (and other validators) are favored over types.
This will let Optional('foo', default=1) take precedence over simply `str` and thus have the opportunity to insert its default.

The only case for which this should change behavior is if somebody was comparing against a type with a `validate` attr. In that case, the logic would have gone down the wrong path anyway, attempting to call validate() on the type itself.