Removed ITERABLE type validating.
There's no need to automatically type check ITERABLE flavours when this can be done by the user. 
The reason this is needed is because, for example, the default JSON encoder accepts both tuples and lists as array types.
Creating a schema with either of them will result in the validation failing if the other is passed instead.