compose uses concept of "reducer"
Defines a "reducer" which will merge schemas together. This operation is performed recursively on the values dictionary schemas. Reducers can be a callable or validator either of which will return a merged schema. The default reducer simply returns the first given schema given in the list of schemas. A callable reducer should have a signature of the form `reducer(schemas)` where `schemas` is a list of schemas. A validator reducer is an object which has a callable attribute "validate" which gets used just like callable reducer.