don't double-format errors. fixes #240
Pull request #238 introduced a change (which was released in 0.7.3) which
caused errors to sometimes be formatted twice, which causeed valid data to
sometimes appear to be invalid, as described in issue #240.

This commit reverts pull request #238, which fixes issue #240, and also adds
error formatting to some other exceptions including the one which gets raised
in the example described as motivation for the change in #238.

It also adds test cases to ensure the behaviors desired in both #238 and #240
remain present.

This commit is also relevant to anyone looking at issue #241 (though it does
not implement the plan described there).